Move url related utility methods to @converse/headless
