+++ package.json
@@ -115,7 +115,6 @@
     "favico.js-slevomat": "^0.3.11",
     "jed": "1.1.1",
     "lit": "^2.0.0-rc.2",
-    "urijs": "^1.19.6",
     "xss": "^1.0.8"
   }
 }
