+++ src/headless/package.json
@@ -48,6 +48,7 @@
     "pluggable.js": "3.0.1",
     "sizzle": "^2.3.5",
     "sprintf-js": "^1.1.2",
-    "strophe.js": "1.4.2"
+    "strophe.js": "1.4.2",
+    "urijs": "^1.19.6"
   }
 }
